start rewriting some of the tests to vitest
