+++ test/util/formatMessage.test.ts
@@ -1,3 +1,4 @@
+import { test, describe, beforeEach, expect } from 'vitest'
 import { printDiffOrStringify, printExpected, printReceived } from 'jest-matcher-utils';
 
 import { enhanceError } from '../../src/util/formatMessage'
