+++ test/utils.test.ts
@@ -1,3 +1,4 @@
+import { describe, test, expect } from 'vitest'
 import { compareNumbers, compareText, compareTextWithArray } from '../src/utils'
 
 describe('utils', () => {
