+++ vitest.config.ts
@@ -1,0 +1,22 @@
+import { defineConfig } from 'vite'
+
+export default defineConfig({
+    test: {
+        include: ['test/**/*.test.ts'],
+        /**
+         * not to ESM ported packages
+         */
+        exclude: [
+            'dist', '.idea', '.git', '.cache',
+            '**/node_modules/**'
+        ],
+        coverage: {
+            enabled: false,
+            exclude: ['**/build/**', '**/__fixtures__/**', '**/*.test.ts'],
+            lines: 96,
+            functions: 90,
+            branches: 94,
+            statements: 96
+        }
+    }
+})
